Remove useless code
